# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cardre v2 API",
    "description": "Auditable open-source credit scorecard builder \u2014 v2.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "description": "Get a single project by ID.",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "artifacts"
        ],
        "summary": "Get Artifact",
        "description": "Get a single artifact by ID.",
        "operationId": "get_artifact_projects__project_id__artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/manual-binning/reviews": {
      "get": {
        "tags": [
          "manual_binning"
        ],
        "summary": "List Reviews",
        "description": "List manual-binning reviews, optionally filtered.",
        "operationId": "list_reviews_projects__project_id__manual_binning_reviews_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "plan_version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Version Id"
            }
          },
          {
            "name": "step_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Step Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualBinningReviewResponse"
                  },
                  "title": "Response List Reviews Projects  Project Id  Manual Binning Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/manual-binning/reviews/{review_id}": {
      "get": {
        "tags": [
          "manual_binning"
        ],
        "summary": "Get Review",
        "description": "Get a single manual-binning review.",
        "operationId": "get_review_projects__project_id__manual_binning_reviews__review_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualBinningReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "manual_binning"
        ],
        "summary": "Update Review",
        "description": "Update a manual-binning review (status, notes).",
        "operationId": "update_review_projects__project_id__manual_binning_reviews__review_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualBinningReviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualBinningReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/manual-binning/edit": {
      "post": {
        "tags": [
          "manual_binning"
        ],
        "summary": "Apply Manual Binning Edit",
        "description": "Apply a manual-binning edit (creates new draft version + review).",
        "operationId": "apply_manual_binning_edit_projects__project_id__manual_binning_edit_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-Project-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Id"
            }
          },
          {
            "name": "X-Project-Path",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Project-Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualBinningEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualBinningEditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/manual-binning/preview": {
      "post": {
        "tags": [
          "manual_binning"
        ],
        "summary": "Preview Binning",
        "description": "Preview WOE/IV/event-rate for a variable's bin data.",
        "operationId": "preview_binning_projects__project_id__manual_binning_preview_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualBinningPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualBinningPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactResponse": {
        "properties": {
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "physical_hash": {
            "type": "string",
            "title": "Physical Hash"
          },
          "logical_hash": {
            "type": "string",
            "title": "Logical Hash"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "artifact_id",
          "artifact_type",
          "role",
          "path",
          "physical_hash",
          "logical_hash",
          "media_type",
          "created_at"
        ],
        "title": "ArtifactResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualBinningEditRequest": {
        "properties": {
          "plan_version_id": {
            "type": "string",
            "title": "Plan Version Id"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "overrides": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Overrides"
          },
          "reviewer_notes": {
            "type": "string",
            "title": "Reviewer Notes",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending"
          },
          "affected_downstream_step_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Downstream Step Ids"
          }
        },
        "type": "object",
        "required": [
          "plan_version_id",
          "step_id"
        ],
        "title": "ManualBinningEditRequest"
      },
      "ManualBinningEditResponse": {
        "properties": {
          "new_plan_version_id": {
            "type": "string",
            "title": "New Plan Version Id"
          },
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "affected_step_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Step Ids"
          }
        },
        "type": "object",
        "required": [
          "new_plan_version_id",
          "review_id"
        ],
        "title": "ManualBinningEditResponse"
      },
      "ManualBinningPreviewRequest": {
        "properties": {
          "variable_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variable Data"
          }
        },
        "type": "object",
        "required": [
          "variable_data"
        ],
        "title": "ManualBinningPreviewRequest"
      },
      "ManualBinningPreviewResponse": {
        "properties": {
          "woe_by_bin": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Woe By Bin"
          },
          "iv": {
            "type": "number",
            "title": "Iv"
          },
          "event_rate_by_bin": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Event Rate By Bin"
          }
        },
        "type": "object",
        "required": [
          "woe_by_bin",
          "iv",
          "event_rate_by_bin"
        ],
        "title": "ManualBinningPreviewResponse"
      },
      "ManualBinningReviewResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "plan_version_id": {
            "type": "string",
            "title": "Plan Version Id"
          },
          "step_id": {
            "type": "string",
            "title": "Step Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_notes": {
            "type": "string",
            "title": "Reviewer Notes",
            "default": ""
          },
          "affected_downstream_step_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Affected Downstream Step Ids"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "plan_version_id",
          "step_id",
          "status"
        ],
        "title": "ManualBinningReviewResponse"
      },
      "ManualBinningReviewUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "reviewer_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Notes"
          }
        },
        "type": "object",
        "title": "ManualBinningReviewUpdate"
      },
...
      "ProjectResponse": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "cardre_version": {
            "type": "string",
            "title": "Cardre Version"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name",
          "created_at",
          "cardre_version"
        ],
        "title": "ProjectResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}